# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Work Pass Extension API",
      "version": "1.0.0"
    },
    "paths": {
      "/api/applications/search": {
        "get": {
          "summary": "Search applications by FIN",
          "parameters": [
            {
              "name": "fin",
              "in": "query",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response"
            },
            "400": {
              "description": "Bad request"
            },
            "404": {
              "description": "Not found"
            },
            "500": {
              "description": "Internal server error"
            }
          }
        }
      },
      "/api/applications/{application_id}/update-expiry": {
        "put": {
          "summary": "Update pass expiry date",
          "parameters": [
            {
              "name": "application_id",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "new_doe": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response"
            },
            "400": {
              "description": "Bad request"
            },
            "404": {
              "description": "Not found"
            }
          }
        }
      },
      "/api/applications/{application_id}/create-stvp": {
        "post": {
          "summary": "Create or extend STVP for expired passes",
          "parameters": [
            {
              "name": "application_id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "STVP extended"
            },
            "201": {
              "description": "New STVP created"
            },
            "400": {
              "description": "Bad request"
            },
            "404": {
              "description": "Not found"
            }
          }
        }
      },
      "/api/applications": {
        "get": {
          "summary": "List all applications",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "type": "integer",
              "default": 1
            },
            {
              "name": "per_page",
              "in": "query",
              "type": "integer",
              "default": 20
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response"
            },
            "404": {
              "description": "Not found"
            }
          }
        }
      },
      "/api/applications/{application_id}/amendments": {
        "get": {
          "summary": "Get amendment history for an application",
          "parameters": [
            {
              "name": "application_id",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "amendment_id": {
                      "type": "string"
                    },
                    "amendment_date": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "original_value": {
                      "type": "string"
                    },
                    "amended_value": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "404": {
              "description": "Not found"
            }
          }
        }
      }
    }
  }
    